initialize(), ArtifactController methods edited for better testing
Whenever we run the app, it would always start with this hardcoded version, so it was like, impossible to see if a change was permanent, actually working and so on. So I added the JSON exports to the ArtifactController, and edited the initialize method like the following. Here it looks at our imported Artifact list first. If this change is a hindrance, please feel free to delete it, we can add it later.